+++ docs/components/content/examples/vue/checkbox/ExampleVueCheckboxIndeterminate.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+const checked = ref<'indeterminate' | boolean>('indeterminate')
+</script>
+
+<template>
+  <NCheckbox
+    v-model:checked="checked"
+    label="Indeterminate"
+  />
+</template>