+++ packages/nuxt/playground/app.vue
@@ -1,9 +1,9 @@
 <template>
   <div class="h-screen w-screen flex flex-col items-center justify-center gap-12 !bg-base">
-    <ColorMode />
+    <NThemeSwitcher />
 
     <Calendar />
 
-    <NThemeSwitcher />
+    <ColorMode />
   </div>
 </template>