+++ packages/nuxt/src/module.ts
@@ -2,15 +2,11 @@ import { addComponentsDir, addImportsDir, addPlugin, createResolver, defineNuxtM
 import { name, version } from '../package.json'
 
 import extendUnocssOptions from './una.config'
-
-interface UnaOptions {
-  primary?: string
-  gray?: string
-}
+import type { UnaSettings } from './runtime/types'
 
 declare module '@nuxt/schema' {
   interface AppConfigInput {
-    una?: UnaOptions
+    una?: UnaSettings
   }
 }
 
@@ -64,6 +60,8 @@ export default defineNuxtModule<ModuleOptions>({
     nuxt.options.appConfig.una = {
       primary: 'yellow',
       gray: 'stone',
+      radius: 0.5,
+      fontSize: 16,
     }
 
     // transpile runtime
@@ -98,6 +96,14 @@ export default defineNuxtModule<ModuleOptions>({
 
     addComponentsDir({
       path: resolve(runtimeDir, 'components/elements', 'dropdown-menu'),
+      prefix: options.prefix,
+      global: options.global,
+      watch: nuxt.options.dev,
+      priority: 10,
+    })
+
+    addComponentsDir({
+      path: resolve(runtimeDir, 'components/elements', 'popover'),
       prefix: options.prefix,
       global: options.global,
       watch: nuxt.options.dev,